+++ src/main/groovy/life/qbic/datamodel/dtos/business/ProductCategoryFactory.groovy
@@ -27,4 +27,20 @@ class ProductCategoryFactory extends EnumFactory<ProductCategory>{
         }
         return desiredKey
     }
+
+    /**
+     * This method returns the enum with the provided String abbreviation.
+     *
+     * @throws IllegalArgumentException in case the String could not be mapped
+     * @param abbreviation the String corresponding to the abbreviation stored in the enum
+     * @return the enum key for the provided abbreviation
+     */
+    ProductCategory getForAbbreviation(String abbreviation) {
+        ProductCategory desiredKey
+        desiredKey = ProductCategory.values().find {it.abbreviation.equals(abbreviation.trim())}
+        if (!desiredKey) {
+            throw new IllegalArgumentException("Unknown abbreviation'$abbreviation' for ${ProductCategory.getSimpleName()}")
+        }
+        return desiredKey
+    }
 }