+++ src/main/groovy/life/qbic/datamodel/dtos/business/services/DataStorage.groovy
@@ -1,6 +1,7 @@
 package life.qbic.datamodel.dtos.business.services
 
 import groovy.transform.EqualsAndHashCode
+import life.qbic.datamodel.dtos.business.ProductCategory
 import life.qbic.datamodel.dtos.business.ProductId
 import life.qbic.datamodel.dtos.business.facilities.Facility
 
@@ -20,12 +21,12 @@ class DataStorage extends PartialProduct {
    * @param description The description of what the product is about.
    * @param unitPrice The price in € per unit
    * @param unit The product unit
-   * @param runningNumber Number used in conjunction with ProductType{@link life.qbic.datamodel.dtos.business.services.ProductType} to identify product
+   * @param runningNumber Number used in conjunction with {@link ProductCategory} to identify product
    * @deprecated since 2.5.0
    */
   @Deprecated
   DataStorage(String name, String description, double unitPrice, ProductUnit unit, String runningNumber) {
-    super(name, description, unitPrice, unit, new ProductId(ProductType.DATA_STORAGE.toString(), runningNumber))
+    super(name, description, unitPrice, unit, new ProductId(ProductCategory.DATA_STORAGE.getAbbreviation(), runningNumber))
   }
 
   /**
@@ -37,13 +38,13 @@ class DataStorage extends PartialProduct {
    * @param description The description of what the product is about.
    * @param unitPrice The price in € per unit
    * @param unit The product unit
-   * @param runningNumber Number used in conjunction with ProductType{@link life.qbic.datamodel.dtos.business.services.ProductType} to identify product
+   * @param runningNumber Number used in conjunction with {@link ProductCategory} to identify product
    *
    * @deprecated 2.11.0
    */
   @Deprecated
   DataStorage(String name, String description, double unitPrice, ProductUnit unit, long runningNumber) {
-    super(name, description, unitPrice, unit, new ProductId.Builder(ProductType.DATA_STORAGE.toString(), runningNumber).build())
+    super(name, description, unitPrice, unit, new ProductId.Builder(ProductCategory.DATA_STORAGE.getAbbreviation(), runningNumber).build())
   }
 
   /**
@@ -56,11 +57,11 @@ class DataStorage extends PartialProduct {
    * @param internalUnitPrice The price in € per unit for internal customers
    * @param externalUnitPrice The price in € per unit for external customers
    * @param unit The product unit
-   * @param runningNumber Number used in conjunction with ProductType{@link life.qbic.datamodel.dtos.business.services.ProductType} to identify product
+   * @param runningNumber Number used in conjunction with {@link ProductCategory} to identify product
    *
    * @since 2.11.0
    */
   DataStorage(String name, String description, double internalUnitPrice, double externalUnitPrice, ProductUnit unit, long runningNumber, Facility serviceProvider) {
-    super(name, description, internalUnitPrice, externalUnitPrice, unit, new ProductId.Builder(ProductType.DATA_STORAGE.toString(), runningNumber).build(), serviceProvider)
+    super(name, description, internalUnitPrice, externalUnitPrice, unit, new ProductId.Builder(ProductCategory.DATA_STORAGE.getAbbreviation(), runningNumber).build(), serviceProvider)
   }
 }